Base for polygonization

+++ src/p_local.h
@@ -271,8 +271,8 @@ fixed_t P_ZAtPoint(const bool_t a_Ceil, const fixed_t a_X, const fixed_t a_Y, co
 extern uint8_t* rejectmatrix;	// for fast sight rejection
 extern size_t g_RJMSize;						// Size of reject matrix
 extern size_t g_BMLSize;		// Block map lump size
-extern long* blockmaplump;		// offsets in blockmap are from here
-extern long* blockmap;			// Big blockmap SSNTails
+extern int32_t* blockmaplump;	// offsets in blockmap are from here
+extern int32_t* blockmap;		// Big blockmap SSNTails
 extern int bmapwidth;
 extern int bmapheight;			// in mapblocks
 extern fixed_t bmaporgx;
